Refactor shared lock logic in tree service.
